refactor: finalize backend conversion logic and modularize sync store
   - move AllFieldsFrame conversion logic to domain-specific impl From blocks
   - split monolithic sync.ts into persistence, hotkeys, and events modules
   - add unit tests for lap_delta computation (Rust) and telemetry formatters (TS)
   - fix linting and formatting issues across the workspace
