+++ src-tauri/src/iracing/frames/mod.rs
@@ -29,8 +29,6 @@ pub use session::SessionFrame;
 use pitwall::{BitField, PitwallFrame};
 use serde::{Deserialize, Serialize};
 
-use super::enums::{SessionState, Skies, TrackSurface};
-
 fn bitfield_to_u32(bits: BitField) -> u32 {
     bits.0
 }
@@ -41,7 +39,7 @@ fn bitfield_to_u32(bits: BitField) -> u32 {
 /// It is private to the `iracing` module and decomposed into domain-specific
 /// frames (`CarDynamicsFrame`, `CarInputsFrame`, etc.) before being emitted
 /// as Tauri events.
-#[derive(PitwallFrame, Serialize, Deserialize, Debug, Clone)]
+#[derive(PitwallFrame, Serialize, Deserialize, Debug, Clone, Default)]
 pub(crate) struct AllFieldsFrame {
     // === Car Dynamics ===
     #[field_name = "Speed"]
@@ -286,182 +284,4 @@ pub(crate) struct AllFieldsFrame {
     pub car_idx_track_surface: Vec<i32>,
     #[field_name = "CarIdxTireCompound"]
     pub car_idx_tire_compound: Vec<i32>,
-}
-
-// === Domain frame decomposition ===
-
-impl From<&AllFieldsFrame> for CarDynamicsFrame {
-    fn from(f: &AllFieldsFrame) -> Self {
-        Self {
-            speed: f.speed,
-            rpm: f.rpm,
-            gear: f.gear,
-            steering_wheel_angle: f.steering_wheel_angle,
-            velocity_x: f.velocity_x,
-            velocity_y: f.velocity_y,
-            velocity_z: f.velocity_z,
-            lat_accel: f.lat_accel,
-            long_accel: f.long_accel,
-            yaw: f.yaw,
-            yaw_rate: f.yaw_rate,
-            pitch: f.pitch,
-            roll: f.roll,
-            shift_indicator_pct: f.shift_indicator_pct,
-            shift_grind_rpm: f.shift_grind_rpm,
-        }
-    }
-}
-
-impl From<&AllFieldsFrame> for CarInputsFrame {
-    fn from(f: &AllFieldsFrame) -> Self {
-        Self {
-            throttle: f.throttle,
-            brake: f.brake,
-            clutch: f.clutch,
-        }
-    }
-}
-
-impl From<&AllFieldsFrame> for CarStatusFrame {
-    fn from(f: &AllFieldsFrame) -> Self {
-        Self {
-            fuel_level: f.fuel_level,
-            fuel_level_pct: f.fuel_level_pct,
-            fuel_use_per_hour: f.fuel_use_per_hour,
-            oil_temp: f.oil_temp,
-            oil_press: f.oil_press,
-            water_temp: f.water_temp,
-            voltage: f.voltage,
-            on_pit_road: f.on_pit_road,
-            is_on_track: f.is_on_track,
-            car_left_right: f.car_left_right,
-            engine_warnings: f.engine_warnings,
-        }
-    }
-}
-
-impl From<&AllFieldsFrame> for LapTimingFrame {
-    fn from(f: &AllFieldsFrame) -> Self {
-        Self {
-            lap: f.lap,
-            lap_dist: f.lap_dist,
-            lap_dist_pct: f.lap_dist_pct,
-            lap_current_lap_time: f.lap_current_lap_time,
-            lap_last_lap_time: f.lap_last_lap_time,
-            lap_best_lap_time: f.lap_best_lap_time,
-            player_car_position: f.player_car_position,
-            player_car_class_position: f.player_car_class_position,
-            lap_delta_to_session_best_live: f.lap_delta_to_session_best_live,
-            lap_delta_to_session_optimal_live: f.lap_delta_to_session_optimal_live,
-        }
-    }
-}
-
-impl From<&AllFieldsFrame> for SessionFrame {
-    fn from(f: &AllFieldsFrame) -> Self {
-        let player_car_flags = f.player_car_idx.and_then(|idx| {
-            f.car_idx_session_flags
-                .get(idx as usize)
-                .map(|bf| bf.0)
-        });
-
-        Self {
-            session_time: f.session_time,
-            session_time_remain: f.session_time_remain,
-            session_state: f.session_state.map(SessionState::from),
-            session_flags: f.session_flags,
-            session_num: f.session_num,
-            session_time_of_day: f.session_time_of_day,
-            player_car_idx: f.player_car_idx,
-            player_car_flags,
-        }
-    }
-}
-
-impl From<&AllFieldsFrame> for CarIdxFrame {
-    fn from(f: &AllFieldsFrame) -> Self {
-        Self {
-            car_idx_lap_dist_pct: f.car_idx_lap_dist_pct.clone(),
-            car_idx_on_pit_road: f.car_idx_on_pit_road.clone(),
-            car_idx_position: f.car_idx_position.clone(),
-            car_idx_class_position: f.car_idx_class_position.clone(),
-            car_idx_lap: f.car_idx_lap.clone(),
-            car_idx_last_lap_time: f.car_idx_last_lap_time.clone(),
-            car_idx_best_lap_time: f.car_idx_best_lap_time.clone(),
-            car_idx_f2_time: f.car_idx_f2_time.clone(),
-            car_idx_est_time: f.car_idx_est_time.clone(),
-            car_idx_track_surface: f
-                .car_idx_track_surface
-                .iter()
-                .map(|&v| TrackSurface::from(v))
-                .collect(),
-            car_idx_tire_compound: f.car_idx_tire_compound.clone(),
-            car_left_right: f.car_left_right,
-        }
-    }
-}
-
-impl From<&AllFieldsFrame> for EnvironmentFrame {
-    fn from(f: &AllFieldsFrame) -> Self {
-        Self {
-            air_temp: f.air_temp,
-            track_temp: f.track_temp,
-            wind_vel: f.wind_vel,
-            wind_dir: f.wind_dir,
-            relative_humidity: f.relative_humidity,
-            skies: f.skies.map(Skies::from),
-            precipitation: f.precipitation,
-            track_wetness: f.track_wetness,
-            weather_declared_wet: f.weather_declared_wet,
-            weather_type: f.weather_type,
-            weather_version: f.weather_version,
-        }
-    }
-}
-
-impl From<&AllFieldsFrame> for ChassisFrame {
-    fn from(f: &AllFieldsFrame) -> Self {
-        Self {
-            lf_ride_height: f.lf_ride_height,
-            rf_ride_height: f.rf_ride_height,
-            lr_ride_height: f.lr_ride_height,
-            rr_ride_height: f.rr_ride_height,
-            lf_shock_defl: f.lf_shock_defl,
-            rf_shock_defl: f.rf_shock_defl,
-            lr_shock_defl: f.lr_shock_defl,
-            rr_shock_defl: f.rr_shock_defl,
-            lf_temp_cl: f.lf_temp_cl,
-            lf_temp_cm: f.lf_temp_cm,
-            lf_temp_cr: f.lf_temp_cr,
-            rf_temp_cl: f.rf_temp_cl,
-            rf_temp_cm: f.rf_temp_cm,
-            rf_temp_cr: f.rf_temp_cr,
-            lr_temp_cl: f.lr_temp_cl,
-            lr_temp_cm: f.lr_temp_cm,
-            lr_temp_cr: f.lr_temp_cr,
-            rr_temp_cl: f.rr_temp_cl,
-            rr_temp_cm: f.rr_temp_cm,
-            rr_temp_cr: f.rr_temp_cr,
-            lf_pressure: f.lf_pressure,
-            rf_pressure: f.rf_pressure,
-            lr_pressure: f.lr_pressure,
-            rr_pressure: f.rr_pressure,
-            lf_wear_l: f.lf_wear_l,
-            lf_wear_m: f.lf_wear_m,
-            lf_wear_r: f.lf_wear_r,
-            rf_wear_l: f.rf_wear_l,
-            rf_wear_m: f.rf_wear_m,
-            rf_wear_r: f.rf_wear_r,
-            lr_wear_l: f.lr_wear_l,
-            lr_wear_m: f.lr_wear_m,
-            lr_wear_r: f.lr_wear_r,
-            rr_wear_l: f.rr_wear_l,
-            rr_wear_m: f.rr_wear_m,
-            rr_wear_r: f.rr_wear_r,
-            lf_brake_temp: f.lf_brake_temp,
-            rf_brake_temp: f.rf_brake_temp,
-            lr_brake_temp: f.lr_brake_temp,
-            rr_brake_temp: f.rr_brake_temp,
-        }
-    }
 }
